WidgetProvider클래스 생성(여러가지 위젯에 대응), 위젯 처리 Worker생성

+++ core/data/src/main/java/io/github/pknujsp/weatherwizard/core/data/widget/WidgetRepository.kt
@@ -1,18 +1,14 @@
 package io.github.pknujsp.weatherwizard.core.data.widget
 
-import io.github.pknujsp.weatherwizard.core.model.EntityModel
-import io.github.pknujsp.weatherwizard.core.model.notification.NotificationEntity
-import io.github.pknujsp.weatherwizard.core.model.notification.ongoing.OngoingNotificationInfoEntity
-import io.github.pknujsp.weatherwizard.core.model.notification.daily.DailyNotificationInfoEntity
 import io.github.pknujsp.weatherwizard.core.model.widget.WidgetEntity
 import kotlinx.coroutines.flow.Flow
 
 interface WidgetRepository {
 
     fun getAll(): Flow<List<WidgetEntity>>
-    suspend fun get(id: Long): WidgetEntity
+    suspend fun get(id: Int): WidgetEntity
 
-    suspend fun add(entity: WidgetEntity): Long
+    suspend fun add(entity: WidgetEntity): Int
 
-    suspend fun delete(id: Long)
+    suspend fun delete(id: Int)
 }
